Split files by class; make SelectionChangedEventArgs constructors internal; Optimizations for SelectionChangedEventArgs; Fire selection changed event/command from property changed handler;
