+++ Xamarin.Forms.Core/Items/SelectionMode.cs
@@ -1,0 +1,9 @@
+﻿namespace Xamarin.Forms
+{
+	public enum SelectionMode 
+	{
+		None,
+		Single,
+		Multiple
+	}
+}
